+++ .changeset/bright-otters-tickle.md
@@ -1,0 +1,9 @@
+---
+'@modern-js/core': minor
+'@modern-js/plugin-analyze': minor
+'@modern-js/types': minor
+---
+
+chore(core): move Hooks types define to `@modern-js/core`
+
+chore(core): 移动 Hooks 类型定义到 `@modern-js/core` 包
