+++ packages/cli/core/src/manager.ts
@@ -12,7 +12,7 @@ import {
   createParallelWorkflow,
 } from '@modern-js/plugin';
 import { compatRequire } from '@modern-js/utils';
-import type { Hooks } from '@modern-js/types';
+import type { Hooks } from './types';
 import type { Command } from './utils/commander';
 import type { NormalizedConfig } from './config/mergeConfig';
 import type { UserConfig } from './config';
