+++ packages/toolkit/types/cli/index.d.ts
@@ -1,5 +1,3 @@
-import { AsyncWaterfall, AsyncWorkflow } from '@modern-js/plugin';
-import { Compiler, MultiCompiler, Configuration } from 'webpack';
 import { ServerRoute } from '../server';
 
 /**
@@ -66,70 +64,4 @@ export interface IAppContext {
   apiOnly: boolean;
   internalDirAlias: string;
   internalSrcAlias: string;
-}
-
-export interface Hooks {
-  beforeDev: AsyncWorkflow<void, unknown>;
-  afterDev: AsyncWorkflow<void, unknown>;
-  beforeCreateCompiler: AsyncWorkflow<
-    {
-      webpackConfigs: Configuration[];
-    },
-    unknown
-  >;
-  afterCreateCompiler: AsyncWorkflow<
-    {
-      compiler: Compiler | MultiCompiler | undefined;
-    },
-    unknown
-  >;
-  beforePrintInstructions: AsyncWaterfall<{
-    instructions: string;
-  }>;
-  beforeBuild: AsyncWorkflow<
-    {
-      webpackConfigs: Configuration[];
-    },
-    unknown
-  >;
-  afterBuild: AsyncWorkflow<void, unknown>;
-  afterMonorepoDeploy: AsyncWorkflow<
-    { operator: any; deployProjectNames: string[] },
-    void
-  >;
-  beforeDeploy: AsyncWorkflow<Record<string, any>, unknown>;
-  afterDeploy: AsyncWorkflow<Record<string, any>, unknown>;
-  modifyEntryExport: AsyncWaterfall<{
-    entrypoint: Entrypoint;
-    exportStatement: string;
-  }>;
-  modifyEntryImports: AsyncWaterfall<{
-    imports: ImportStatement[];
-    entrypoint: Entrypoint;
-  }>;
-  modifyEntryRuntimePlugins: AsyncWaterfall<{
-    entrypoint: Entrypoint;
-    plugins: RuntimePlugin[];
-  }>;
-  modifyEntryRenderFunction: AsyncWaterfall<{
-    entrypoint: Entrypoint;
-    code: string;
-  }>;
-  modifyFileSystemRoutes: AsyncWaterfall<{
-    entrypoint: Entrypoint;
-    routes: Route[];
-  }>;
-  modifyServerRoutes: AsyncWaterfall<{
-    routes: ServerRoute[];
-  }>;
-  htmlPartials: AsyncWaterfall<{
-    entrypoint: Entrypoint;
-    partials: HtmlPartials;
-  }>;
-  addRuntimeExports: AsyncWaterfall<void>;
-  beforeGenerateRoutes: AsyncWaterfall<{
-    entrypoint: Entrypoint;
-    code: string;
-  }>;
-  addDefineTypes: AsyncWaterfall<void>;
 }
