+++ packages/toolkit/types/package.json
@@ -13,10 +13,6 @@
   ],
   "version": "1.5.6",
   "types": "./index.d.ts",
-  "dependencies": {
-    "@modern-js/plugin": "workspace:^1.4.2",
-    "webpack": "^5.54.0"
-  },
   "exports": {
     ".": "./index.d.ts",
     "./server": {
